backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pymongo
motor
lxml